list/renderer: fix viewport adjustment on PageDown
With the fix from #387, when PageDown moved the cursor below the
viewport, the renderer broke early before rendering the focused item, so
`selectedLineStart` and `selectedLineEnd` stayed -1, which leads to the
viewport not being adjusted and rerendered.

Now the renderer continues rendering until it reaches the focused item,
even if it's below the viewport. So that the viewport is adjusted to
show the focused item and the screen scrolls to follow the cursor